Add ability to specify or deactivate CC
- Leaving CC args blank will keep default CCs
- Passing in arg 6 & 7 will use the new CCs
- Putting -1 as arg will deactivate CC

Doesn't work well if the same channel is used. The problem is the CCs are sent one after another, and the region can't tell if the touch is in the region until both CCs are received. As a result, when the first CC of an outside touch is received it's possible to still think it's in the region, and will trigger one of the CCs to change for the old region.